% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_anatogram.R
\name{get_anatogram}
\alias{get_anatogram}
\title{Get anatogram from defaults in gganatogram}
\usage{
get_anatogram(organism = "human", sex = "male", anatogram = NULL)
}
\arguments{
\item{organism}{The organism to be plotted.  Currently,
only \code{human} is accepted.}

\item{sex}{Sex of the organism}

\item{anatogram}{A list, similar to \code{\link{hgMale_list}}
that will create the outline and has the corresponding organ
\code{data.frame}s in that list}
}
\value{
A list of values for the \code{anatogram}
}
\description{
Get anatogram from defaults in gganatogram
}
\examples{
get_anatogram()
get_anatogram(sex = "female")
get_anatogram(organism = "mouse")
}
