% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gganatogram.R
\name{gganatogram}
\alias{gganatogram}
\title{gganatogram}
\usage{
gganatogram(data = NULL, outline = TRUE, fillOutline = "lightgray",
  organism = "human", sex = "male", fill = "colour")
}
\arguments{
\item{data}{The main data frame consisting of what organs to plot, colours, and values. Default is NULL}

\item{fillOutline}{Fill colour of outline. Default is #a6bddb}

\item{organism.}{Species to plot. Default is human.}
}
\description{
This function plots anatograms of specified tissues, species, and sex .
}
\examples{

#First lets just plot the outline. Only male human is possible now
gganatogram(fillOutline='#a6bddb', organism='human', sex='male', fill="colour")


#To add organs, create a data frame with specified tissues


organPlot <- data.frame(organ = c("heart", "leukocyte", "nerve", "brain",
     "liver", "stomach", "colon"),
 type = c("circulation", "circulation",
     "nervous system", "nervous system", "digestion", "digestion",
     "digestion"),
 colour = c("red", "red", "purple", "purple", "orange",
     "orange", "orange"),
 value = c(10, 5, 1, 8, 2, 5, 5),
 stringsAsFactors=F)


gganatogram(data=organPlot, fillOutline='#a6bddb', organism='human', sex='male', fill="colour")


#We can also remove the outline
gganatogram(data=organPlot, outline=FALSE, fillOutline='#a6bddb', organism='human', sex='male', fill="colour")


organPlot \%>\%
     dplyr::filter(type \%in\% 'circulation') \%>\%
 gganatogram(fillOutline='#a6bddb', organism='human', sex='male', fill="colour")

organPlot \%>\%
     dplyr::filter(type \%in\% c('circulation', 'nervous system')) \%>\%
 gganatogram(fillOutline='#a6bddb', organism='human', sex='male', fill="value") +
 theme_void() +
 scale_fill_gradient(low = "white", high = "red")

gganatogram(data=oraganPlot, fillOutline='#a6bddb', organism='human', sex='male', fill="colour") +
facet_wrap(~type)

#Use hgMale_key to find all tissues to plot
hgMale_key = gganatogram::hgMale_key
head(hgMale_key)
gganatogram(data=hgMale_key, fillOutline='#a6bddb', organism='human', sex='male', fill="colour") + theme_void()


gganatogram(data=hgMale_key, fillOutline='#a6bddb', organism='human', sex='male', fill="colour") + theme_void() + facet_wrap(~type, ncol=3)


gganatogram(data=hgMale_key, fillOutline='#a6bddb', organism='human', sex='male', fill="colour")


gganatogram(data=hgMale_key, fillOutline='#a6bddb', organism='human', sex='male', fill="value")


gganatogram(data=hgMale_key, fillOutline='#a6bddb', organism='human', sex='male', fill="colour") +facet_wrap(~type, ncol=3) +void()


}
\keyword{anatogram}
\keyword{organs}
\keyword{tissues}
