% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgFemale_key.R
\docType{data}
\name{hgFemale_key}
\alias{hgFemale_key}
\title{Color Key for Human Female}
\format{A \code{data.frame} with 4 columns, which are:
\describe{
\item{organ}{the name of the organ}
\item{type}{organ system: circulation, digestion, nervous
system, other, respiratory }
\item{colour}{color to highlight of the data}
\item{value}{the value to fill the data in an example}
}}
\usage{
hgFemale_key
}
\description{
A \code{data.frame} with the data to plot the human female.
}
\keyword{datasets}
