% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmMale_key.R
\docType{data}
\name{mmMale_list}
\alias{mmMale_list}
\title{Plotting Coordinates for Mouse Male}
\format{A \code{list} with 340 elements columns, each of which are
a \code{data.frame} with 6 columns:
\describe{
\item{X1}{coordinates in x-dimension}
\item{X2}{coordinates in y-dimension}
\item{id}{identifier of the region}
\item{x}{coordinates in x-dimension, duplicated for ease of use}
\item{y}{coordinates in x-dimension, duplicated for ease of use}
\item{group}{polygon group to avoid linkage between polygons }
}}
\usage{
mmMale_list
}
\description{
A \code{list} with the coordinate data to plot the Mouse male.
}
\keyword{datasets}
